+++ onsiteprint-plugin/CHANGELOG.md
@@ -90,6 +90,7 @@ This document includes the change log notes for the OnsitePrint Plugin - [GitHub
 - Changed the `Import` of the `Participant`.
 - Added the `Create Participant Listener` Function.
 - Added the `Column Input Listener` Function.
+- Added the `Download CSV File Listener` Function.
 
     [EB | participant-listeners.js](blocks/event/block-script-parts/participant-listeners.js)